Cleanup TimespanList implementations
- Extract `to_lilypond` and `explode` to helper modules
- Remove `sorted_into_non_overlapping_sublists` and use `explode` instead
- Refactor Offset validation for `split`
